requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
jinja2>=3.1.0
assemblyai
murf

google-generativeai>=0.5.0
